Removed un-used routes, moved selectable attributes component to presentation preparation module, clean up un-used components
