+++ src/app/app-routing.module.ts
@@ -5,26 +5,18 @@ import {NavigateService} from '@core/services/navigate.service';
 import {VerifierEndpointService} from "@core/services/verifier-endpoint.service";
 
 const routes: Routes = [
-	{ path: '', redirectTo: 'home', pathMatch: 'full' },
+	{ path: '',
+    redirectTo: 'home',
+    pathMatch: 'full' },
 	{ path: 'home', loadComponent: () => import('@features/presentation-request-preparation/home/home.component')
       .then(c => c.HomeComponent) },
 	{ path: 'custom-request',
     loadChildren: () => import('@features/custom-presentation-request/custom-presentation-request.module').
       then(m => m.CustomPresentationRequestModule )},
-	{ path: 'selectable',
-		loadChildren: () => import('@features/selectable-presentation/selectable-presentation.module').
-			then(m => m.SelectablePresentationModule )
-  },
-	{ path: 'invoke-wallet',
+  { path: 'invoke-wallet',
     loadChildren: () => import('@features/invoke-wallet/invoke-wallet.module').
-			then(m => m.InvokeWalletModule)
-	},
-	{ path: 'mdl-selectable',
-    loadChildren: () => import('@features/selectable-presentation/selectable-presentation.module').
-      then(m => m.SelectablePresentationModule ),
-	},
-	{
-		path: 'get-wallet-code',
+			then(m => m.InvokeWalletModule) },
+	{ path: 'get-wallet-code',
 		loadComponent: () => import('./features/wallet-redirect/wallet-redirect.component').then(c => c.WalletRedirectComponent),
 		providers: [VerifierEndpointService, NavigateService],
 		resolve: {
