+++ src/app/features/presentation-request-preparation/components/selectable-attestation-attributes/model/DialogResult.ts
