+++ src/app/features/presentation-request-preparation/components/selectable-attestation-attributes/selectable-attestation-attributes.component.html
